pyproject.toml
# pyproject.toml

[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "airtable_pg_sync"
version = "0.0.37"
description = "Sync Airtable bases to a Postgres schemas in real time"
readme = "README.md"
authors = [{ name = "Benjamin Urwin", email = "benurwin@outlook.com" }]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["airtable", "postgres", "sync", "realtime", "webhook"]
dependencies = ["requests", "psycopg[binary]", "aiohttp", "click", "pyyaml", "rich", "python-dateutil"]
requires-python = ">=3.9"

[project.urls]
Homepage = "https://github.com/benurwin/airtable_pg_sync"

[project.scripts]
airtable_pg_sync = "airtable_pg_sync:main"
